fix: inconsistent regen:approvedMethodologies fields
